+++ modules/nio/src/main/java/java/nio/ReadWriteFloatArrayBuffer.java
@@ -104,7 +104,7 @@ final class ReadWriteFloatArrayBuffer extends FloatArrayBuffer {
 	}
 
     public FloatBuffer put(float[] src, int off, int len) {
-        if (off < 0 || len < 0 || off + len > src.length) {
+        if (off < 0 || len < 0 || (long)off + (long)len > src.length) {
             throw new IndexOutOfBoundsException();
         }
         if (len > remaining()) {
